% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStats.R
\name{grouped2_summary_stats}
\alias{grouped2_summary_stats}
\title{Summary Stats with 2 grouping variables}
\usage{
grouped2_summary_stats(data, group1, group2, target_var, decimals = 2)
}
\arguments{
\item{data}{A data frame}

\item{group1}{A character representing a grouping variable}

\item{group2}{A character representing a grouping variable}

\item{target_var}{An outcome variable}

\item{decimals}{Number of decimals in the output frame}
}
\value{
A data frame with min, mean, median, sd, max
}
\description{
It works with data.frame, data.table and tibble.
}
\examples{

library(rlang)
library(tidyverse)
library(glue)
library(summaryStatsR)
library(datasets)

grouped2_summary_stats(datasets::mtcars, group1 = "gear", group2 = "cyl", target_var='mpg')

}
