% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStats.R
\name{grouped_summary_stats}
\alias{grouped_summary_stats}
\title{Summary Stats with a grouping variable
It works with data.frame, data.table and tibble.}
\usage{
grouped_summary_stats(data, group, target_var, decimals = 2)
}
\arguments{
\item{data}{A data frame}

\item{group}{A character representing a grouping variable}

\item{target_var}{A character representingthe outcome variable}

\item{decimals}{Number of decimals in the output frame}
}
\value{
A data frame with min, mean, median, sd, max
}
\description{
Summary Stats with a grouping variable
It works with data.frame, data.table and tibble.
}
\examples{

library(rlang)
library(dplyr)
library(purrr)
library(glue)
library(summaryStatsR)
library(datasets)

grouped_summary_stats(datasets::iris, group = "Species", target_var = "Sepal.Length")

}
