% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{Calculate Standard Error of the Meam (SEM)}
\usage{
sem(x)
}
\arguments{
\item{x}{Input Vector}
}
\value{
Standard Error of the Mean. A single value
}
\description{
Calculate Standard Error of the Meam (SEM)
}
\examples{

library(summaryStatsR)

sem(iris[['Petal.Length']])

}
