% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStats.R
\name{summary_stats}
\alias{summary_stats}
\title{A Simple Custom-defined Summary Stats Function for a single variable
It works with data.frame, data.table and tibble.}
\usage{
summary_stats(data, target_var, decimals = 2)
}
\arguments{
\item{data}{A data frame}

\item{target_var}{A character representing the outcome variable}

\item{decimals}{Number of decimals in the output frame}
}
\value{
A data frame with min, mean, median, sd, max
}
\description{
A Simple Custom-defined Summary Stats Function for a single variable
It works with data.frame, data.table and tibble.
}
\examples{

library(rlang)
library(dplyr)
library(purrr)
library(glue)
library(summaryStatsR)
library(datasets)

summary_stats(datasets::iris, "Sepal.Length", decimals = 2)
}
